requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
Django==2.0.6
Flask==1.0.2
Flask-Cors==3.0.4
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.2.0
pluggy==0.6.0
py==1.5.3
pytest==3.6.1
pytz==2018.4
PyYAML==5.1.2
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
